server/requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai==1.52.0
requests==2.31.0
beautifulsoup4==4.12.2
numpy==1.24.3
scikit-learn==1.3.0
python-dotenv==1.0.0
pydantic==2.5.0
